requirements.txt
# Core
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.15,<4.0

# Filtering
django-filter>=24.1,<25.0

# PostgreSQL driver
psycopg2-binary>=2.9,<3.0

# API Documentation (Swagger / OpenAPI)
drf-spectacular>=0.27,<1.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0

# Linting & Formatting
black>=24.0,<25.0
ruff>=0.3,<1.0
